fix: no need to have public key inside config
