+++ config.ts
@@ -32,8 +32,6 @@ export const DEFAULT_FIREBLOCKS_API_SECRET_KEY =
   'resources/keys/fireblocks-priv.pem';
 export const DEFAULT_DFNS_SERVICE_ACCOUNT_PRIVATE_KEY =
   'resources/keys/dfns-priv.pem';
-export const DEFAULT_AWS_KMS_PUBLIC_KEY =
-  'resources/keys/aws-public-b6961cc4-ca88-4bcf-9e0f-51696c7fd230.pem';
 export const TEST_TIMEOUT = 10000; // 10 seconds
 
 //* Fireblocks configuration
@@ -79,41 +77,9 @@ export const dfnsConfig_ECDSA = new DFNSConfig(
 );
 
 //* AWS KMS configuration
-const kmsPubKeyPathOrEncoded =
-  process.env.AWS_KMS_PUBLIC_KEY || DEFAULT_AWS_KMS_PUBLIC_KEY;
-let kmsPubKeyPem: string | undefined;
-let kmsPubKeyDerHex: string | undefined;
-if (kmsPubKeyPathOrEncoded) {
-  if (
-    pathRegex.test(kmsPubKeyPathOrEncoded) &&
-    fs.existsSync(kmsPubKeyPathOrEncoded)
-  ) {
-    kmsPubKeyPem = fs.readFileSync(
-      path.resolve(kmsPubKeyPathOrEncoded),
-      'utf8'
-    );
-  } else if (!pathRegex.test(kmsPubKeyPathOrEncoded)) {
-    kmsPubKeyPem = Buffer.from(kmsPubKeyPathOrEncoded, 'base64').toString(
-      'utf8'
-    );
-  } else {
-    kmsPubKeyPem = undefined;
-  }
-  if (kmsPubKeyPem) {
-    const kmsPubKeyDerBuffer = Buffer.from(
-      kmsPubKeyPem.replace(
-        /^-----BEGIN PUBLIC KEY-----[\r\n]+|[\r\n]+-----END PUBLIC KEY-----$/g,
-        ''
-      ),
-      'base64'
-    );
-    kmsPubKeyDerHex = kmsPubKeyDerBuffer.toString('hex');
-  }
-}
 export const awsKMSConfig = new AWSKMSConfig(
   process.env.AWS_ACCESS_KEY_ID ?? '',
   process.env.AWS_SECRET_ACCESS_KEY ?? '',
   process.env.AWS_REGION ?? '',
-  process.env.AWS_KMS_KEY_ID ?? '',
-  kmsPubKeyDerHex ?? ''
+  process.env.AWS_KMS_KEY_ID ?? ''
 );
