+++ examples/index.ts
@@ -26,6 +26,7 @@ import {
   dfnsConfig_ECDSA,
   fireblocksConfig,
   awsKMSConfig,
+  pathRegex,
 } from '../config';
 import { AWSKMSConfig, DFNSConfig, FireblocksConfig } from '../src';
 import { WriteStream, readFileSync } from 'fs';
@@ -36,6 +37,9 @@ import KeyListExample from './KeyListExample';
 import HscsExample from './HscsExample';
 import HfssExample from './HfssExample';
 import HcsExample from './HcsExample';
+import fs from 'fs';
+import path from 'path';
+import { PublicKey } from '@hashgraph/sdk';
 
 async function main(): Promise<void> {
   console.log('👋 Welcome to the Hedera Custodians Integration example');
@@ -197,39 +201,76 @@ async function main(): Promise<void> {
       break;
 
     case 'AWS Key Management Service':
+      // Get the AWS KMS public key
+      const DEFAULT_AWS_KMS_PUBLIC_KEY =
+        'resources/keys/aws-public-b6961cc4-ca88-4bcf-9e0f-51696c7fd230.pem';
+      const kmsPubKeyPathOrEncoded =
+        process.env.AWS_KMS_PUBLIC_KEY || DEFAULT_AWS_KMS_PUBLIC_KEY;
+      let kmsPubKeyPem: string | undefined;
+      let kmsPublicKey: PublicKey | undefined;
+      if (kmsPubKeyPathOrEncoded) {
+        if (
+          pathRegex.test(kmsPubKeyPathOrEncoded) &&
+          fs.existsSync(kmsPubKeyPathOrEncoded)
+        ) {
+          kmsPubKeyPem = fs.readFileSync(
+            path.resolve(kmsPubKeyPathOrEncoded),
+            'utf8'
+          );
+        } else if (!pathRegex.test(kmsPubKeyPathOrEncoded)) {
+          kmsPubKeyPem = Buffer.from(kmsPubKeyPathOrEncoded, 'base64').toString(
+            'utf8'
+          );
+        } else {
+          kmsPubKeyPem = undefined;
+        }
+        if (kmsPubKeyPem) {
+          const kmsPubKeyDerBuffer = Buffer.from(
+            kmsPubKeyPem.replace(
+              /^-----BEGIN PUBLIC KEY-----[\r\n]+|[\r\n]+-----END PUBLIC KEY-----$/g,
+              ''
+            ),
+            'base64'
+          );
+          const kmsPubKeyDerHex = kmsPubKeyDerBuffer.toString('hex');
+          kmsPublicKey = PublicKey.fromString(kmsPubKeyDerHex);
+        }
+      }
       if (custodialAnwsers.useEnvVars) {
         example = new Example(
           awsKMSConfig,
           process.env.AWS_KMS_HEDERA_ACCOUNT_ID ?? '',
-          awsKMSConfig.awsKmsPublicKey
+          kmsPublicKey ?? ''
         );
         htsExample = new HtsExample(
           awsKMSConfig,
           process.env.AWS_KMS_HEDERA_ACCOUNT_ID ?? '',
-          awsKMSConfig.awsKmsPublicKey
+          kmsPublicKey ?? ''
         );
         hscsExample = new HscsExample(
           awsKMSConfig,
           process.env.AWS_KMS_HEDERA_ACCOUNT_ID ?? '',
-          awsKMSConfig.awsKmsPublicKey
+          kmsPublicKey ?? ''
         );
         hfssExample = new HfssExample(
           awsKMSConfig,
           process.env.AWS_KMS_HEDERA_ACCOUNT_ID ?? '',
-          awsKMSConfig.awsKmsPublicKey
+          kmsPublicKey ?? ''
         );
         hcsExample = new HcsExample(
           awsKMSConfig,
           process.env.AWS_KMS_HEDERA_ACCOUNT_ID ?? '',
-          awsKMSConfig.awsKmsPublicKey
+          kmsPublicKey ?? ''
         );
         keyListExample = new KeyListExample(
           awsKMSConfig,
           process.env.AWS_KMS_HEDERA_ACCOUNT_ID ?? '',
-          awsKMSConfig.awsKmsPublicKey
+          kmsPublicKey ?? ''
         );
       } else {
-        const awsKmsParams = await askAwsKmsParams();
+        const awsKmsParams = await askAwsKmsParams({
+          kmsPubKeyPem: kmsPubKeyPem,
+        });
         example = new Example(awsKmsParams);
         htsExample = new HtsExample(awsKmsParams);
         hscsExample = new HscsExample(awsKmsParams);
@@ -439,7 +480,9 @@ async function askFireblocksParams(): Promise<ExampleConfig> {
   );
 }
 
-async function askAwsKmsParams(): Promise<ExampleConfig> {
+async function askAwsKmsParams({
+  kmsPubKeyPem,
+}: { kmsPubKeyPem?: string } = {}): Promise<ExampleConfig> {
   const awsKmsParams: Answers = await inquirer.prompt([
     {
       type: 'input',
@@ -469,7 +512,7 @@ async function askAwsKmsParams(): Promise<ExampleConfig> {
       type: 'input',
       name: 'publicKey',
       message: 'Enter the AWS KMS public key',
-      default: awsKMSConfig.awsKmsPublicKey,
+      default: kmsPubKeyPem,
     },
     {
       type: 'input',
@@ -478,16 +521,28 @@ async function askAwsKmsParams(): Promise<ExampleConfig> {
       default: process.env.AWS_KMS_HEDERA_ACCOUNT_ID ?? '',
     },
   ]);
+  // Parse the AWS KMS public key from PEM to PublicKey
+  if (!kmsPubKeyPem) {
+    throw new Error('❌ Missing AWS KMS public key');
+  }
+  const kmsPubKeyDerBuffer = Buffer.from(
+    kmsPubKeyPem.replace(
+      /^-----BEGIN PUBLIC KEY-----[\r\n]+|[\r\n]+-----END PUBLIC KEY-----$/g,
+      ''
+    ),
+    'base64'
+  );
+  const kmsPubKeyDerHex = kmsPubKeyDerBuffer.toString('hex');
+  const kmsPublicKey = PublicKey.fromString(kmsPubKeyDerHex);
   return new ExampleConfig(
     new AWSKMSConfig(
       awsKmsParams.accessKeyId,
       awsKmsParams.secretAccessKey,
       awsKmsParams.region,
-      awsKmsParams.kmsKeyId,
-      awsKmsParams.publicKey
+      awsKmsParams.kmsKeyId
     ),
     awsKmsParams.hederaAccountId,
-    awsKmsParams.publicKey
+    kmsPublicKey
   );
 }
